// lib/types.ts
export interface Skip {
  id: string;
  name: string;
  size: string;
  price: number;
  hirePeriod: string;
  imageUrl?: string;
  notAllowedOnRoad: boolean;
}

export interface ApiSkip {
  id: number;
  size: number;
  hire_period_days: number;
  transport_cost: number | null;
  per_tonne_cost: number | null;
  price_before_vat: number;
  vat: number;
  postcode: string;
  area: string;
  forbidden: boolean;
  created_at: string;
  updated_at: string;
  allowed_on_road: boolean;
  allows_heavy_waste: boolean;
}